+++ jwt_allauth/login/views.py
@@ -1,4 +1,3 @@
-from django.conf import settings
 from rest_framework import status
 from rest_framework.request import Request
 from rest_framework.response import Response
@@ -6,14 +5,15 @@ from rest_framework_simplejwt.exceptions import TokenError, InvalidToken
 from rest_framework_simplejwt.views import TokenObtainPairView
 from rest_framework.throttling import AnonRateThrottle
 
-from jwt_allauth.app_settings import LoginSerializer
+from jwt_allauth import app_settings
 from jwt_allauth.utils import get_user_agent, sensitive_post_parameters_m, build_token_response
-from jwt_allauth.constants import REFRESH_TOKEN_COOKIE
 
 
 class LoginView(TokenObtainPairView):
-    serializer_class = LoginSerializer
     throttle_classes = [AnonRateThrottle]
+
+    def get_serializer_class(self):
+        return app_settings.LoginSerializer
 
     @sensitive_post_parameters_m
     def dispatch(self, *args, **kwargs):